reorganize interfaces
