+++ internal/interfaces/function_definition.go
@@ -1,0 +1,9 @@
+package interfaces
+
+import "github.com/Kolterdyx/mcbasic/internal/types"
+
+type FunctionDefinition struct {
+	Name       string
+	Args       []TypedIdentifier
+	ReturnType types.ValueType
+}
