+++ internal/interfaces/typed_identifier.go
@@ -1,0 +1,8 @@
+package interfaces
+
+import "github.com/Kolterdyx/mcbasic/internal/types"
+
+type TypedIdentifier struct {
+	Name string
+	Type types.ValueType
+}
